Refactor collapse-budget-groups Changed the name to be toggle-master-categories. The user setting is retained and mapped to the old value of collapseExpandBudgetGroups.
