+++ sauce/features/budget/toggle-master-categories/settings.js
@@ -1,0 +1,8 @@
+module.exports = {
+  name: 'ToggleMasterCategories',
+  type: 'checkbox',
+  default: false,
+  section: 'budget',
+  title: 'Toggle All Master Categories Open/Close',
+  description: 'Adds a button to the Budget Toolbar to open or close all master categories at once.'
+};
