--- source/common/res/features/collapse-budget-groups/settings.json
@@ -1,14 +1,0 @@
-{
-         "name": "collapseExpandBudgetGroups",
-         "type": "checkbox",
-      "default": false,
-      "section": "budget",
-        "title": "Add Expand/Collapse budget master categories button",
-  "description": "Adds a button that allows you expand and collapse all budget groups.",
-      "actions": {
-                    "true": [
-                      "injectCSS", "main.css",
-                      "injectScript", "main.js"
-                    ]
-                 }
-}
